chore(dashboard): remove unused onboarding wizard and webgl stack
