+++ client/dashboard/package.json
@@ -51,9 +51,6 @@
     "@radix-ui/react-tabs": "^1.1.15",
     "@radix-ui/react-toggle-group": "^1.1.13",
     "@radix-ui/react-tooltip": "^1.2.11",
-    "@react-three/drei": "^10.7.7",
-    "@react-three/fiber": "^9.6.1",
-    "@react-three/postprocessing": "^3.0.4",
     "@speakeasy-api/moonshine": "1.44.0",
     "@tailwindcss/vite": "catalog:",
     "@tanstack/react-query": "catalog:",
@@ -77,14 +74,12 @@
     "next-themes": "^0.4.6",
     "nuqs": "^2.9.0",
     "posthog-js": "^1.393.4",
-    "postprocessing": "^6.39.1",
     "radix-ui": "^1.6.0",
     "react": "^19.2.7",
     "react-chartjs-2": "^5.3.1",
     "react-dom": "^19.2.7",
     "react-error-boundary": "^6.1.2",
     "react-markdown": "^10.1.0",
-    "react-merge-refs": "^3.0.2",
     "react-router": "^8.2.0",
     "recharts": "^2.15.4",
     "remark-gfm": "^4.0.1",
@@ -93,8 +88,6 @@
     "streamdown": "^1.6.11",
     "tailwind-merge": "^3.6.0",
     "tailwindcss": "^4.3.2",
-    "three": "^0.181.2",
-    "tunnel-rat": "^0.1.2",
     "tw-animate-css": "^1.4.0",
     "tw-shimmer": "^0.4.11",
     "use-stick-to-bottom": "^1.1.6",
@@ -113,7 +106,6 @@
     "@types/node": "^24.13.2",
     "@types/react": "catalog:",
     "@types/react-dom": "catalog:",
-    "@types/three": "^0.181.0",
     "@vitejs/plugin-react": "catalog:",
     "happy-dom": "^20.10.6",
     "knip": "^6.20.0",
