+++ client/dashboard/src/App.css
@@ -40,14 +40,6 @@
   font-style: normal;
   font-display: swap;
 }
-@font-face {
-  font-family: "speakeasyAscii";
-  src: url("./fonts/speakeasy/speakeasy-ascii.woff2") format("woff2");
-  font-weight: 400;
-  font-style: normal;
-  font-display: swap;
-}
-
 :root {
   --optional-label: "optional";
