--- client/dashboard/src/components/webgl/constants.ts
@@ -1,1 +1,0 @@
-export const CANVAS_PADDING = 0.25;
