--- client/dashboard/src/components/webgl/index.tsx
@@ -1,3 +1,0 @@
-export { WebGLCanvas } from "./canvas";
-export { FontTexture } from "./components/ascii-effect/font-texture";
-export { AsciiVideo } from "./ascii-video";
