--- client/dashboard/src/components/webgl/tunnel.tsx
@@ -1,19 +1,0 @@
-import { Fragment, useId } from "react";
-import tunnel from "tunnel-rat";
-
-const WebGL = tunnel();
-
-export const WebGLIn = ({
-  children,
-}: {
-  children: React.ReactNode;
-}): JSX.Element => {
-  const id = useId();
-  return (
-    <WebGL.In>
-      <Fragment key={id}>{children}</Fragment>
-    </WebGL.In>
-  );
-};
-
-export const WebGLOut = WebGL.Out as () => React.ReactNode;
