--- client/dashboard/src/lib/webgl/utils.ts
@@ -1,1 +1,0 @@
-export const glsl = (x: TemplateStringsArray): string => x[0]!;
