+++ client/dashboard/src/pages/onboarding/UploadOpenAPI.tsx
@@ -8,11 +8,9 @@ import UploadAssetStep from "@/components/upload-asset/step";
 import UploadAssetStepper from "@/components/upload-asset/stepper";
 import { useStepper } from "@/components/upload-asset/stepper/use-stepper";
 import UploadFileStep from "@/components/upload-asset/upload-file-step";
-import { cn } from "@/lib/utils";
 import { useRoutes } from "@/routes";
-import { useDeploymentLogs } from "@gram/client/react-query/deploymentLogs.js";
 import { Heading } from "@/components/ui/heading";
-import { Button, CodeSnippet, Stack } from "@speakeasy-api/moonshine";
+import { Button, Stack } from "@speakeasy-api/moonshine";
 import { ArrowRightIcon, FileTextIcon, RefreshCcwIcon } from "lucide-react";
 
 export default function UploadOpenAPI(): JSX.Element {
@@ -139,54 +137,4 @@ function FooterActions() {
         </>
       );
   }
-}
-
-export function DeploymentLogs(props: {
-  deploymentId: string;
-  onlyErrors?: boolean;
-}): JSX.Element | null {
-  const { data, status, error } = useDeploymentLogs({
-    deploymentId: props.deploymentId,
-  });
-
-  if (status === "pending") {
-    return <Type>Loading deployment logs...</Type>;
-  }
-
-  if (status === "error") {
-    return (
-      <div>
-        <Type>Error loading deployment logs</Type>
-        <CodeSnippet code={error.toString()} language="text" />
-      </div>
-    );
-  }
-
-  if (data == null) {
-    return null;
-  }
-
-  const lines = (
-    props.onlyErrors
-      ? data.events.filter((e) => e.event.includes("error"))
-      : data.events
-  ).map((e) => {
-    return (
-      <p
-        key={e.id}
-        className={cn(
-          e.event.includes("error") && "text-destructive",
-          "rounded px-4 py-1 hover:bg-gray-100 dark:hover:bg-white/15",
-        )}
-      >
-        {e.message}
-      </p>
-    );
-  });
-
-  return (
-    <div className="max-h-[250px] overflow-y-auto font-mono text-sm">
-      {lines.length > 0 ? lines : "OpenAPI document processed without issue"}
-    </div>
-  );
 }
